refactor(sweep): externalize SweepForm
